028.implement-strstr/problem28.go
// Implement strStr().

// Return the index of the first occurrence of needle in haystack, or -1 if needle is not part of haystack.

// Example 1:

// Input: haystack = "hello", needle = "ll"
// Output: 2
// Example 2:

// Input: haystack = "aaaaa", needle = "bba"
// Output: -1

package leetcode

func strStr(haystack string, needle string) int {
	if needle == "" {
		return 0
	}
	for i := 0; i+len(needle)-1 < len(haystack); i++ {
		if haystack[i:i+len(needle)] == needle {
			return i
		}
	}
	return -1
}
